requirements.txt
mcp
google-generativeai
gradio
python-dotenv
google-adk
